+++ src/Unit/Demon.h
@@ -10,7 +10,7 @@ class Demon : public Soldier {
         bool superAttackIsReady;
 
     public:
-        Demon(const char* name, int maxHp = 100, int damage = 10);
+        Demon(const char* name, float maxHp = 100, float damage = 10);
 
         void superAttack(Unit* enemy);
         void description();
