+++ src/Unit/Soldier.cpp
@@ -2,7 +2,7 @@
 
 #define DEBUG 1
 
-Soldier::Soldier(const char* name, int maxHp, int damage) {
+Soldier::Soldier(const char* name, float maxHp, float damage) {
     if ( DEBUG ) {
         std::cout << "CONSTRUCTOR SOLDIER" << std::endl;
     }
